Fix wrong import when working with akka
Previously, we would not check if the file renames were all working with an actual symbol. This was not a problem previously, but recently we added a check that would always try to use renamed imports, which would break in case of a NoSymbol. Now, we jsut check it before adding to the renames map.

Fixes https://github.com/scalameta/metals/issues/2736
